+++ resources/release-notes.md
@@ -1,6 +1,6 @@
 # App
 
-- Add support for dark mode in mission selection background
+- Add support for dark mode in toolbar icons and mission selection background
 - Add checkbox in IMAP AT/SFT visualization options to show spectrograms
 - Reduce duplication in definition of supported view-controllers for each mission
 - Disable toolbar when selecting a mission